fastApi/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
billiard==4.2.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
kombu==5.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mutagen==1.47.0
orjson==3.10.3
prompt_toolkit==3.0.47
pycryptodomex==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
requests==2.32.3
rich==13.7.1
rq==1.16.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
yt-dlp==2024.5.27